+++ app/routers/userAuthentication/auth.py
@@ -4,7 +4,7 @@ from starlette import status
 
 from app.models.userModel import SignupRequest, User
 from app.routers.userAuthentication.security import create_access_token, get_current_user, create_refresh_token
-from app.database.userDatabase import create_user, authenticate_user
+from app.database.database import create_user, authenticate_user
 
 authRouter = APIRouter()
